pyproject.toml
[project]
name = "levin-qa-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.97.0",
    "pdfplumber>=0.11.7",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.3",
    "pypdf>=5.8.0",
    "python-dotenv>=1.1.1",
]
